Add support for loading referenced schemas from a proxy.
When a schema is loaded through a proxy that doesn't correspond to the
id of the schema, then the referenced schemas will be loaded from the
same proxy.
This very usefull when schema's are being developed and/or are not
hosted yet on the final server where the id is pointing at.
This can easily be 'hacked' into the current code by replacing the fully
qualified url's of the missing references by relative url's.
The remaining of the code will then automatically use the dowload url of
the referencing schema to resolve the relative url intead of using the
id to resolve the relative url.